[PF-1540] Update Sam & TCL clients -- WIP (#837)
* PF-1965: When cloning resources, clear certain properties (a.k.a TERRA_WORKSPACE_FOLDER_ID) (#826)

* clear terra_workspace_folder_id key on cloning

* spotless

* add comment

* add comment

* spotless

* when destination workspace is the same as source, do not clear the property

* PF-1984: add folder properties (#832)

* add folder properties

* fix a check

* address comments

* bump 0.254.392

* [PF-1930] Fix dismal failure with CreateStorageTransferServiceJobStep undo (#833)

* [PF-1930] Fix dismal failure with CreateStorageTransferServiceJobStep undo

* fix test

* beforeAll afterAll

* call workspaceService.deleteWorkspace and delete WorkspaceConnectedTestUtils.deleteWorkspaceAndCloudContext

* bump 0.254.393

* changes to root project

* Java 17 support

+++ service/src/main/java/bio/terra/workspace/service/resource/controlled/flight/clone/bucket/CreateStorageTransferServiceJobStep.java
@@ -140,6 +140,31 @@ public final class CreateStorageTransferServiceJobStep implements Step {
   // previous step's undo method.
   @Override
   public StepResult undoStep(FlightContext flightContext) throws InterruptedException {
+    final FlightMap workingMap = flightContext.getWorkingMap();
+    final String transferJobName =
+        workingMap.get(ControlledResourceKeys.STORAGE_TRANSFER_JOB_NAME, String.class);
+    final String controlPlaneProjectId =
+        workingMap.get(ControlledResourceKeys.CONTROL_PLANE_PROJECT_ID, String.class);
+
+    TransferJob transferJob;
+    try {
+      transferJob =
+          StorageTransferServiceUtils.getTransferJob(
+              storagetransfer, transferJobName, controlPlaneProjectId);
+    } catch (IOException e) {
+      return new StepResult(StepStatus.STEP_RESULT_FAILURE_FATAL, e);
+    }
+
+    // transferJob == null means transfer job doesn't exist, and can happen if:
+    // - doStep() failed before transfer job was created
+    // - This undoStep() has run before
+    // transferJob.getStatus() == "DELETED" can happen if:
+    // - DeleteStorageTransferServiceJobStep.doStep() has succeeded. (Don't try to delete job again;
+    //   that will result in dismal failure.)
+    if (transferJob == null || transferJob.getStatus().equals("DELETED")) {
+      return StepResult.getStepResultSuccess();
+    }
+
     // A failure to delete will result in a DISMAL_FAILURE of the flight.
     return StorageTransferServiceUtils.deleteTransferJobStepImpl(flightContext, storagetransfer);
   }
